test: AppConfig 의존성 주입
Before:
- 구현체에 관한 테스트만을 유지함.

After:
- 구현체 테스트와 역할 테스트를 구분
- AppConfig 설정을 불러와 테스트하는 Application 환경에 맞는 테스트 코드 추가
